+++ .eslintignore
@@ -1,2 +1,3 @@
 dist
 webpack.config.js
+renderTodoList.js
